+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ## Added
+- support for theme embedding within stylesheets
 - `LiveViewNative.Jetpack.normalize_os_version/1`
 - `LiveViewNative.Jetpack.normalize_app_version/1`
 - CoreComponents